feat: 스킬값을 0.1 단위 소수로 변경 + 조커 삽입 에러 로깅
- SKILLS를 소수값으로 변경 (1.0~5.0, 0.6 간격)
- 모든 팀 평균이 정확히 0.1 단위로 계산됨
- 조커 skill_value를 4.4(세미프로)로 변경
- 조커 참가자 INSERT 에러 로깅 추가
- DB 스키마 skill_value를 NUMERIC(3,1)로 변경

+++ tournament-app/supabase/schema.sql
@@ -22,7 +22,7 @@ CREATE TABLE IF NOT EXISTS participants (
   tournament_id UUID REFERENCES tournaments(id) ON DELETE CASCADE,
   nickname VARCHAR(20) NOT NULL,
   skill VARCHAR(10) NOT NULL,
-  skill_value INTEGER NOT NULL CHECK (skill_value BETWEEN 1 AND 5),
+  skill_value NUMERIC(3,1) NOT NULL CHECK (skill_value BETWEEN 1.0 AND 5.0),
   card_tier VARCHAR(10) DEFAULT '브론즈',
   profile_image_url TEXT,
   card_image_url TEXT,
